Refactor UI and data models for step 2.
