+++ src/cplus_plugin/gui/model_component_widget.py
@@ -22,9 +22,11 @@ from .component_item_model import (
     NcsPathwayItemModel,
     NCS_PATHWAY_TYPE,
 )
+from ..conf import settings_manager
 from .implementation_model_editor_dialog import ImplementationModelEditorDialog
 from .ncs_pathway_editor_dialog import NcsPathwayEditorDialog
-from ..models.base import ImplementationModel, LayerType, NcsPathway
+from ..models.base import ImplementationModel, NcsPathway
+from ..utils import FileUtils
 
 
 WidgetUi, _ = loadUiType(
@@ -43,19 +45,24 @@ class ModelComponentWidget(QtWidgets.QWidget, WidgetUi):
         if self._item_model is not None:
             self.item_model = self._item_model
 
-        add_icon = QgsApplication.instance().getThemeIcon("symbologyAdd.svg")
+        add_icon = FileUtils.get_icon("symbologyAdd.svg")
         self.btn_add.setIcon(add_icon)
         self.btn_add.clicked.connect(self._on_add_item)
 
-        remove_icon = QgsApplication.instance().getThemeIcon("symbologyRemove.svg")
+        remove_icon = FileUtils.get_icon("symbologyRemove.svg")
         self.btn_remove.setIcon(remove_icon)
         self.btn_remove.setEnabled(False)
         self.btn_remove.clicked.connect(self._on_remove_item)
 
-        edit_icon = QgsApplication.instance().getThemeIcon("mActionToggleEditing.svg")
+        edit_icon = FileUtils.get_icon("mActionToggleEditing.svg")
         self.btn_edit.setIcon(edit_icon)
         self.btn_edit.setEnabled(False)
         self.btn_edit.clicked.connect(self._on_edit_item)
+
+        reload_icon = QgsApplication.instance().getThemeIcon("mActionReload.svg")
+        self.btn_reload.setIcon(reload_icon)
+        self.btn_reload.setToolTip(self.tr("Refresh view"))
+        self.btn_reload.clicked.connect(self._on_reload)
 
     @property
     def item_model(self) -> ComponentItemModelType:
@@ -97,6 +104,10 @@ class ModelComponentWidget(QtWidgets.QWidget, WidgetUi):
         :type text: str
         """
         self.lbl_title.setText(text)
+
+    def load(self):
+        """Subclass to determine how to initialize the items."""
+        pass
 
     def _on_add_item(self):
         """Slot raised when add item button has been clicked.
@@ -181,6 +192,18 @@ class ModelComponentWidget(QtWidgets.QWidget, WidgetUi):
 
         return [self._item_model.item(idx.row()) for idx in idxs]
 
+    def clear(self):
+        """Remove all items in the view. To be implemented
+        by subclasses."""
+        pass
+
+    def _on_reload(self):
+        """Slot raised when the reload button has been clicked.
+        Default implementation is to call the load the items
+        afresh.
+        """
+        self.load()
+
 
 class NcsComponentWidget(ModelComponentWidget):
     """Widget for displaying and managing NCS pathways."""
@@ -206,6 +229,12 @@ class NcsComponentWidget(ModelComponentWidget):
         :type ncs_pathway: NcsPathway
         """
         self.item_model.add_ncs_pathway(ncs_pathway)
+
+    def clear(self):
+        """Removes all NCS pathway items in the view."""
+        items = self.ncs_items()
+        for item in items:
+            self.item_model.remove_ncs_pathway(item.uuid)
 
     def _update_ui_on_selection_changed(self):
         """Temporarily disable edit, remove buttons."""
@@ -279,6 +308,14 @@ class NcsComponentWidget(ModelComponentWidget):
             self.item_model.remove_ncs_pathway(str(ncs.uuid))
             self.clear_description()
 
+    def load(self):
+        """Load items from settings."""
+        self.clear()
+
+        ncs_pathways = settings_manager.get_all_ncs_pathways()
+        for ncs in ncs_pathways:
+            self.add_ncs_pathway(ncs)
+
 
 class ImplementationModelComponentWidget(ModelComponentWidget):
     """Widget for displaying and managing implementation models."""
@@ -291,6 +328,8 @@ class ImplementationModelComponentWidget(ModelComponentWidget):
         self.lst_model_items.setAcceptDrops(True)
         self.lst_model_items.setDragDropMode(QtWidgets.QAbstractItemView.DropOnly)
         self.lst_model_items.setDropIndicatorShown(True)
+
+        self.btn_reload.setVisible(False)
 
     def models(self) -> typing.List[ImplementationModel]:
         """Returns a collection of ImplementationModel objects in the
@@ -311,6 +350,20 @@ class ImplementationModelComponentWidget(ModelComponentWidget):
         :rtype: list
         """
         return self.item_model.model_items()
+
+    def clear(self):
+        """Removes all implementation model items in the view."""
+        items = self.model_items()
+        for item in items:
+            self.item_model.remove_implementation_model(item.uuid)
+
+    def load(self):
+        """Load implementation models from settings."""
+        self.clear()
+
+        imp_models = settings_manager.get_all_implementation_models()
+        for imp_model in imp_models:
+            self.add_implementation_model(imp_model)
 
     def _on_add_item(self):
         """Show implementation model editor."""
