+++ src/cplus_plugin/utils.py
@@ -5,11 +5,19 @@
 
 
 import os
+from pathlib import Path
 
 from qgis.PyQt import QtCore, QtGui
 from qgis.core import Qgis, QgsApplication, QgsMessageLog
 
-from .definitions.defaults import DOCUMENTATION_SITE
+from .conf import settings_manager
+from .definitions.constants import NCS_PATHWAY_SEGMENT
+from .definitions.defaults import (
+    DEFAULT_IMPLEMENTATION_MODELS,
+    DEFAULT_NCS_PATHWAYS,
+    DOCUMENTATION_SITE,
+)
+from .definitions.constants import NCS_PATHWAY_SEGMENT
 
 
 def tr(message):
@@ -65,6 +73,59 @@ def open_documentation(url=None):
     return result
 
 
+def is_dark_theme() -> bool:
+    """Checks if the current QGIS UI theme is dark mode.
+
+    :returns: True if the current UI theme is on dark mode, else False.
+    :rtype: bool
+    """
+    if QgsApplication.instance().themeName() == "Night Mapping":
+        return True
+
+    return False
+
+
+def initialize_default_settings():
+    """Initialize default model components such as NCS pathways
+    and implementation models.
+
+    It will check if there are existing components using the UUID
+    and only add the ones that do not exist in the settings.
+
+    This is normally called during plugin startup.
+    """
+    # Add default pathways
+    for ncs_dict in DEFAULT_NCS_PATHWAYS:
+        try:
+            ncs_uuid = ncs_dict["uuid"]
+            ncs = settings_manager.get_ncs_pathway(ncs_uuid)
+            if ncs is None:
+                # Update dir
+                # base_dir = settings_manager.get_value(Settings.BASE_DIR, None)
+                base_dir = None
+                if base_dir is not None:
+                    file_name = ncs_dict["path"]
+                    absolute_path = f"{base_dir}/{NCS_PATHWAY_SEGMENT}/{file_name}"
+                    abs_path = str(os.path.normpath(absolute_path))
+                    ncs_dict["path"] = abs_path
+                ncs_dict["user_defined"] = False
+                settings_manager.save_ncs_pathway(ncs_dict)
+        except KeyError as ke:
+            log(f"Default NCS configuration load error - {str(ke)}")
+            continue
+
+    # Add default implementation models
+    for imp_model_dict in DEFAULT_IMPLEMENTATION_MODELS:
+        try:
+            imp_model_uuid = imp_model_dict["uuid"]
+            imp_model = settings_manager.get_implementation_model(imp_model_uuid)
+            if imp_model is None:
+                settings_manager.save_implementation_model(imp_model_dict)
+        except KeyError as ke:
+            log(f"Default implementation model configuration load error - {str(ke)}")
+            continue
+
+
 class FileUtils:
     """
     Provides functionality for commonly used file-related operations.
@@ -96,14 +157,12 @@ class FileUtils:
 
         return QtGui.QIcon(icon_path)
 
-
-def is_dark_theme() -> bool:
-    """Checks if the current QGIS UI theme is dark mode.
-
-    :returns: True if the current UI theme is on dark mode, else False.
-    :rtype: bool
-    """
-    if QgsApplication.instance().themeName() == "Night Mapping":
-        return True
-
-    return False
+    @staticmethod
+    def create_ncs_pathways_dir(base_dir: str):
+        """Creates an NCS sub-directory under BASE_DIR. Skips
+        creation of the sub-directory if it already exists.
+        """
+        ncs_pathway_dir = f"{base_dir}/{NCS_PATHWAY_SEGMENT}"
+        p = Path(ncs_pathway_dir)
+        if not p.exists():
+            p.mkdir(parents=True)
